apps/engine/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-binance
pandas
httpx
redis
psycopg2-binary
python-dotenv
apscheduler
dateparser
